requirements.txt
requests>=2.32.0
beautifulsoup4>=4.12.3
pandas>=2.2.2
python-dateutil>=2.9.0.post0
python-dotenv>=1.0.0
msal>=1.24.0
